Squashed 'data_compression/' changes from 349650a..2e6d899
git-subtree-dir: data_compression
git-subtree-split: 2e6d899d03b2d64d51954697fff5aa6c08115ac0

+++ L2/demos/gzip/src/host.cpp
@@ -179,7 +179,7 @@ void xil_compress_top(std::string& compress_mod, std::string& single_bin, uint8_
         std::cout << "Unable to open file";
         exit(1);
     }
-    uint32_t input_size = get_file_size(inFile);
+    uint64_t input_size = get_file_size(inFile);
 
     const char* sizes[] = {"B", "kB", "MB", "GB", "TB"};
     double len = input_size;
